reorganise components file structure, fix header text clip
