+++ src/App.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import About from './components/About';
-import Header from './components/Header';
+import Header from './components/common/Header';
 import MeshCanvas from './components/MeshCanvas';
 import BackgroundCanvas from './components/BackgroundCanvas';
-import ContactFormModal from './components/modal/ContactForm/ContactFormModal';
+import ContactFormModal from './components/ContactForm/ContactFormModal';
 
 const App: React.FC = () => {
   return (
@@ -15,7 +15,7 @@ const App: React.FC = () => {
 
       <div className="min-h-screen flex flex-col lg:flex-row items-center justify-center w-full">
 
-        {/* Header section (visible on medium screens and up) */}
+        {/* Page url at top left of the page (visible on medium screens and up) */}
         <div className="absolute top-0 left-0 p-5 hidden md:block">
           <Header text="trsctr.github.io" textSize="text-2xl" className="tracking-widest font-mono z-20" hasGradient />
         </div> 
@@ -25,13 +25,7 @@ const App: React.FC = () => {
             
             {/* Textbox section */}
             <div className="flex-1 p-5 z-30">
-            <About title="Hello, my name is Otto" imageUrl="/assets/photo.jpg" imageOnRight>
-              <p className="mb-2">I'm an aspiring software developer with a background in audio engineering and tech support, passionate about creative coding.</p>
-              <p>I recently completed my core studies at <a href="http://hive.fi" className="text-primary hover:text-accent transition-all" target="_blank">Hive Helsinki</a>, where I gained hands-on experience in programming with C and C++. </p>
-              <p className="mt-2">I built this page so I could have some fun with React, Tailwind, and Three.js, and showcase my projects in the future.</p>
-              <p className="mt-2">Hope you like what you see! Feel free to interact with that mesh, check out my work and connect with me on LinkedIn or GitHub.</p>
-              <p className="mt-2">Photo by <a href="https://www.instagram.com/diego_j_videos_and_photography/" className="text-primary hover:text-accent transition-all" target="_blank">Diego James</a>.</p>
-            </About>
+            <About/>
             <ContactFormModal/>
             </div>
