+++ src/components/common/FormField.tsx
@@ -96,7 +96,7 @@ const FormField: React.FC<FormFieldProps> = ({
                         peer-placeholder-shown:translate-y-0 peer-focus:scale-75
                         peer-focus:-translate-y-6"
             >
-               {label}
+            {label}
             </label>
         </div>
     );
